refactor(AuthenticatorFactory): Get Authenticator instance based on auth type
